PR fixes :)

+++ src/app/api/file/route.ts
@@ -1,41 +1,15 @@
 import { NextResponse } from "next/server";
 import { File, FileResponse } from "./route.schema";
-import { google } from "googleapis";
 import { prisma } from "@server/db/client";
-import { withSession } from "../../../server/decorator";
-import { env } from "../../../env/server.mjs";
-
-const initializeDriveAuth = async (request) => {
-  const { access_token, refresh_token } = (await prisma.account.findFirst({
-    where: {
-      userId: request.session.user.id,
-    },
-  })) ?? { access_token: null };
-
-  const auth = new google.auth.OAuth2({
-    clientId: env.GOOGLE_CLIENT_ID,
-    clientSecret: env.GOOGLE_CLIENT_SECRET,
-  });
-
-  auth.setCredentials({
-    access_token,
-    refresh_token,
-  });
-
-  const service = google.drive({
-    version: "v3",
-    auth,
-  });
-
-  const body = await request.req.json();
-  const fileRequest = File.safeParse(body);
-
-  return { service, fileRequest };
-};
+import { withSession } from "@server/decorator";
+import { createDriveService } from "@server/service";
 
 export const POST = withSession(async (request) => {
   try {
-    const { service, fileRequest } = await initializeDriveAuth(request);
+    const service = await createDriveService(request.session.user.id);
+
+    const body = await request.req.json();
+    const fileRequest = File.safeParse(body);
 
     if (!fileRequest.success) {
       console.log(fileRequest.error);
@@ -50,14 +24,25 @@ export const POST = withSession(async (request) => {
       const fileData = fileRequest.data;
 
       // Check that user has this senior assigned to them
-      const { SeniorIDs } = await prisma.user.findFirst({
+      const user = await prisma.user.findFirst({
         where: {
           id: request.session.user.id,
         },
       });
 
+      if (user === null || user.SeniorIDs === null) {
+        return NextResponse.json(
+          FileResponse.parse({
+            code: "NO_USER",
+            message: "User does not exist",
+          })
+        );
+      }
+
       if (
-        !SeniorIDs.some((seniorId: string) => seniorId === fileData.seniorId)
+        !user.SeniorIDs.some(
+          (seniorId: string) => seniorId === fileData.seniorId
+        )
       ) {
         return NextResponse.json(
           FileResponse.parse({
@@ -82,7 +67,8 @@ export const POST = withSession(async (request) => {
         );
       }
 
-      const parentID = foundSenior.folder.split("/").pop();
+      const parentID = foundSenior.folder;
+
       const formatted_date =
         (fileData.date.getMonth() > 8
           ? fileData.date.getMonth() + 1
@@ -113,7 +99,7 @@ export const POST = withSession(async (request) => {
       const googleFileId = file.data.id;
 
       // If the data is valid, save it to the database via prisma client
-      await prisma?.file.create({
+      await prisma.file.create({
         data: {
           date: fileData.date,
           filetype: fileData.filetype,
@@ -145,7 +131,10 @@ export const POST = withSession(async (request) => {
 
 export const PATCH = withSession(async (request) => {
   try {
-    const { service, fileRequest } = await initializeDriveAuth(request);
+    const service = await createDriveService(request.session.user.id);
+
+    const body = await request.req.json();
+    const fileRequest = File.safeParse(body);
 
     if (!fileRequest.success) {
       console.log(fileRequest.error);
@@ -160,19 +149,30 @@ export const PATCH = withSession(async (request) => {
       const fileData = fileRequest.data;
 
       // Check that user has this senior assigned to them
-      const { SeniorIDs } = await prisma.user.findFirst({
+      const user = await prisma.user.findFirst({
         where: {
           id: request.session.user.id,
         },
       });
 
+      if (user === null || user.SeniorIDs === null) {
+        return NextResponse.json(
+          FileResponse.parse({
+            code: "NO_USER",
+            message: "User does not exist",
+          })
+        );
+      }
+
       if (
-        !SeniorIDs.some((seniorId: string) => seniorId === fileData.seniorId)
+        !user.SeniorIDs.some(
+          (seniorId: string) => seniorId === fileData.seniorId
+        )
       ) {
         return NextResponse.json(
           FileResponse.parse({
-            code: "NOT_AUTHORIZED",
-            message: "Senior not assigned to user",
+            code: "NO_SENIOR",
+            message: "Senior does not exist",
           }),
           { status: 404 }
         );
@@ -222,14 +222,24 @@ export const PATCH = withSession(async (request) => {
           fileUpdateData
         );
 
-        const { id } = await prisma.file.findFirst({
+        const file = await prisma.file.findFirst({
           where: {
             url: fileData.url,
           },
         });
 
+        if (file == null) {
+          return NextResponse.json(
+            FileResponse.parse({
+              code: "NO_FILE",
+              message: "File does not exist",
+            }),
+            { status: 404 }
+          );
+        }
+
         await prisma.file.update({
-          where: { id: id },
+          where: { id: file.id },
           data: { date: fileData.date, Tags: fileData.Tags },
         });
 
@@ -264,7 +274,10 @@ export const PATCH = withSession(async (request) => {
 
 export const DELETE = withSession(async (request) => {
   try {
-    const { service, fileRequest } = await initializeDriveAuth(request);
+    const service = await createDriveService(request.session.user.id);
+
+    const body = await request.req.json();
+    const fileRequest = File.safeParse(body);
 
     if (!fileRequest.success) {
       console.log(fileRequest.error);
@@ -279,14 +292,25 @@ export const DELETE = withSession(async (request) => {
       const fileData = fileRequest.data;
 
       // Check that user has this senior assigned to them
-      const { SeniorIDs } = await prisma.user.findFirst({
+      const user = await prisma.user.findFirst({
         where: {
           id: request.session.user.id,
         },
       });
 
+      if (user === null || user.SeniorIDs === null) {
+        return NextResponse.json(
+          FileResponse.parse({
+            code: "NO_USER",
+            message: "User does not exist",
+          })
+        );
+      }
+
       if (
-        !SeniorIDs.some((seniorId: string) => seniorId === fileData.seniorId)
+        !user.SeniorIDs.some(
+          (seniorId: string) => seniorId === fileData.seniorId
+        )
       ) {
         return NextResponse.json(
           FileResponse.parse({
@@ -322,14 +346,24 @@ export const DELETE = withSession(async (request) => {
           fileId: googleFileId,
         });
 
-        const { id } = await prisma.file.findFirst({
+        const file = await prisma.file.findFirst({
           where: {
             url: fileData.url,
           },
         });
 
+        if (file == null) {
+          return NextResponse.json(
+            FileResponse.parse({
+              code: "NO_FILE",
+              message: "File does not exist",
+            }),
+            { status: 404 }
+          );
+        }
+
         await prisma.file.delete({
-          where: { id: id },
+          where: { id: file.id },
         });
 
         return NextResponse.json(
